.local/share/meme-tagger/pyproject.toml
[project]
name = "meme-tagger"
version = "0.1.0"
description = "Local-VLM image tagging bot. Telegram + CLI ingest, sidecar-JSON output, grep/jq retrieval."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-telegram-bot>=21.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "Pillow>=10.0.0",
    "av>=12.0.0",
    "rapidocr-onnxruntime>=1.3.0",
]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100

[tool.uv]
package = false
